registry_dashboard/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pyjwt==2.9.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
python-multipart==0.0.18
pandas==2.2.3
openpyxl==3.1.5
